move skipForward to be a factory for Schedules
rather than just being a function in TaskUtils, create a more
generic function that advances a schedule ahead by the required number
of iterations to a given date, optionally returning a new schedule with
the remaining recurrences and new startdate.
