+++ src/main/scala/org/apache/mesos/chronos/schedule/Schedule.scala
@@ -1,5 +1,5 @@
 package org.apache.mesos.chronos.schedule
 
 abstract class Schedule {
- 
+
 }
